+++ previous-versions/accounting/pkg/models/operations/isagedcreditorsreportavailable.go
@@ -3,6 +3,7 @@
 package operations
 
 import (
+	"github.com/codatio/client-sdk-go/previous-versions/accounting/pkg/models/shared"
 	"net/http"
 )
 
@@ -21,6 +22,8 @@ func (o *IsAgedCreditorsReportAvailableRequest) GetCompanyID() string {
 type IsAgedCreditorsReportAvailableResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
+	// Your API request was not properly authorized.
+	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -34,6 +37,13 @@ func (o *IsAgedCreditorsReportAvailableResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
+}
+
+func (o *IsAgedCreditorsReportAvailableResponse) GetErrorMessage() *shared.ErrorMessage {
+	if o == nil {
+		return nil
+	}
+	return o.ErrorMessage
 }
 
 func (o *IsAgedCreditorsReportAvailableResponse) GetStatusCode() int {