+++ previous-versions/accounting/pkg/models/shared/companydataset.go
@@ -147,13 +147,13 @@ func (o *CompanyDatasetWeblink) GetURL() *string {
 	return o.URL
 }
 
-// CompanyDataset - > View the coverage for company info in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=cashFlowStatement" target="_blank">Data coverage explorer</a>.
+// CompanyDataset - > View the coverage for company profile in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=cashFlowStatement" target="_blank">Data coverage explorer</a>.
 //
 // Company info provides standard details about a linked company such as their address, phone number, and company registration.
 //
 // > **Company information or companies?**
 // >
-// > Company information is standard information that is held in the accounting platform about a company. `Companies` is an endpoint that lists businesses in the Codat system that have linked and shared their data sources.
+// > Company profile is standard information that is held in the accounting platform about a company. `Companies` is an endpoint that lists businesses in the Codat system that have linked and shared their data sources.
 type CompanyDataset struct {
 	// Identifier or reference for the company in the accounting platform.
 	AccountingPlatformRef *string `json:"accountingPlatformRef,omitempty"`